--- docs/suspensive.org/src/content/en/docs/codemods/migrateSuspensiveReactQueryPackage.mdx
@@ -1,57 +1,0 @@
----
-description: 'A codemod that migrates the removed @suspensive/react-query package to the version-specific @suspensive/react-query-4 or @suspensive/react-query-5.'
----
-
-import { Callout } from '@/components'
-
-# Migrate `@suspensive/react-query` package
-
-<Callout type='info'>
-
-- Recommended when updating from `@suspensive/react-query` v3 to `@suspensive/react-query` v4.
-- To also convert APIs deprecated in Suspensive React Query to those provided by TanStack Query, please use the [`tanstack-query-import`](./tanstackQueryImport) Codemod.
-
-</Callout>
-
-In v4, the unified `@suspensive/react-query` package has been removed. You need to switch to the version-specific package (`@suspensive/react-query-4` or `@suspensive/react-query-5`) that matches the installed major version of `@tanstack/react-query`.
-
-```bash filename="Terminal"
-npx @suspensive/codemods migrate-suspensive-react-query-package .
-```
-
-For example:
-
-```tsx /@suspensive/react-query/
-import {
-  useSuspenseQuery,
-  SuspenseQuery,
-  queryOptions,
-  IsFetching,
-} from '@suspensive/react-query'
-import * as RQ from '@suspensive/react-query'
-export { queryOptions, QueryClientConsumer } from '@suspensive/react-query'
-```
-
-Transforms into:
-
-```tsx /@tanstack/react-query/ /@suspensive/react-query-5/
-import { useSuspenseQuery, queryOptions } from '@tanstack/react-query'
-import { SuspenseQuery, IsFetching } from '@suspensive/react-query-5'
-import * as RQ from '@suspensive/react-query-5'
-export { queryOptions } from '@tanstack/react-query'
-export { QueryClientConsumer } from '@suspensive/react-query-5'
-```
-
-## Version detection
-
-Starting from the current working directory and walking upward, it checks the nearest `package.json` to determine whether the project depends on v4 or v5 of `@tanstack/react-query`. The target package (`@suspensive/react-query-4` vs `@suspensive/react-query-5`) is determined based on the detected major version.
-
-If detection fails, it falls back to v5 and prints a warning. To specify the target explicitly, set the `SUSPENSIVE_RQ_TARGET` environment variable to `4` or `5` before running:
-
-```bash filename="Terminal"
-SUSPENSIVE_RQ_TARGET=4 npx @suspensive/codemods migrate-suspensive-react-query-package .
-```
-
-## Notes
-
-- It does not modify `package.json`. After running, remove `@suspensive/react-query` from your dependencies and make sure either `@suspensive/react-query-4` or `@suspensive/react-query-5` is installed.
