+++ docs/suspensive.org/src/content/en/docs/codemods/migrateToV4.mdx
@@ -1,24 +1,22 @@
 ---
-description: 'One-shot codemod to handle every code-level breaking change introduced in @suspensive/react-query v4.'
+description: 'A codemod for handling breaking changes in @suspensive/react-query v4.'
 ---
 
 import { Callout } from '@/components'
 
-# Migrate to v4
+# Migrating to v4
 
 <Callout type='info'>
 
-- Recommended when upgrading `@suspensive/react-query` from v3 to v4 — runs every code-level migration in a single command.
-- Updating your `package.json` dependencies is still a manual step.
+- Recommended when updating from `@suspensive/react-query` v3 to v4.
+- This codemod does not modify `package.json`. After running it, remove `@suspensive/react-query` from your dependencies and make sure `@suspensive/react-query-4` or `@suspensive/react-query-5` is installed.
 
 </Callout>
 
-`@suspensive/react-query` v4 contains two code-level breaking changes:
+There are two code-level breaking changes in `@suspensive/react-query` v4:
 
-1. The unified `@suspensive/react-query` package is removed. Imports must point at the version-specific `@suspensive/react-query-4` or `@suspensive/react-query-5`.
-2. APIs backported into `@tanstack/react-query` (`useSuspenseQuery`, `queryOptions`, `mutationOptions`, `usePrefetchInfiniteQuery`, and so on) are removed from `@suspensive/react-query-{4,5}`. They must be imported directly from `@tanstack/react-query`.
-
-Internally chains [`migrate-suspensive-react-query-package`](./migrateSuspensiveReactQueryPackage) followed by [`tanstack-query-import`](./tanstackQueryImport).
+1. The unified package `@suspensive/react-query` has been removed. All imports must point to the version-specific package `@suspensive/react-query-4` or `@suspensive/react-query-5`.
+2. APIs provided by `@tanstack/react-query` (such as `queryOptions`, `mutationOptions`, etc.) are also removed from `@suspensive/react-query-{4,5}`. These must be imported directly from `@tanstack/react-query`.
 
 ```bash filename="Terminal"
 npx @suspensive/codemods migrate-to-v4 .
@@ -26,14 +24,16 @@ npx @suspensive/codemods migrate-to-v4 .
 
 For example:
 
-```tsx /@suspensive/react-query/
+```tsx /@suspensive/react-query/ /@suspensive/react-query-5/
 import {
   useSuspenseQuery,
   SuspenseQuery,
   queryOptions,
   IsFetching,
 } from '@suspensive/react-query'
-import { mutationOptions } from '@suspensive/react-query-5'
+import { queryOptions, PrefetchQuery } from '@suspensive/react-query-5'
+import * as RQ from '@suspensive/react-query'
+export { queryOptions, QueryClientConsumer } from '@suspensive/react-query'
 ```
 
 Transforms into:
@@ -41,14 +41,18 @@ Transforms into:
 ```tsx /@tanstack/react-query/ /@suspensive/react-query-5/
 import { useSuspenseQuery, queryOptions } from '@tanstack/react-query'
 import { SuspenseQuery, IsFetching } from '@suspensive/react-query-5'
-import { mutationOptions } from '@tanstack/react-query'
+import { queryOptions } from '@tanstack/react-query'
+import { PrefetchQuery } from '@suspensive/react-query-5'
+import * as RQ from '@suspensive/react-query-5'
+export { queryOptions } from '@tanstack/react-query'
+export { QueryClientConsumer } from '@suspensive/react-query-5'
 ```
 
 ## Version detection
 
-The codemod walks up from the current working directory and inspects the nearest `package.json` to determine whether you depend on `@tanstack/react-query` v4 or v5. The detected major version selects the destination package (`@suspensive/react-query-4` vs `@suspensive/react-query-5`).
+Starting from the current working directory and walking upward, the codemod locates the nearest `package.json` to determine whether the project depends on v4 or v5 of `@tanstack/react-query`. The detected major version is used when rewriting `@suspensive/react-query` to its version-specific source.
 
-If detection fails, the codemod falls back to **v5** with a warning. To force a target explicitly, set the `SUSPENSIVE_RQ_TARGET` environment variable to `4` or `5`:
+If detection fails, the codemod falls back to v5 and prints a warning. To specify the target explicitly, set the `SUSPENSIVE_RQ_TARGET` environment variable to `4` or `5` before running:
 
 ```bash filename="Terminal"
 SUSPENSIVE_RQ_TARGET=4 npx @suspensive/codemods migrate-to-v4 .
@@ -56,4 +60,4 @@ SUSPENSIVE_RQ_TARGET=4 npx @suspensive/codemods migrate-to-v4 .
 
 ## Notes
 
-- The codemod does **not** update your `package.json`. After running it, remove `@suspensive/react-query` from your dependencies and ensure `@suspensive/react-query-4` or `@suspensive/react-query-5` is installed.
+- This codemod does not modify `package.json`. After running it, remove `@suspensive/react-query` from your dependencies and make sure `@suspensive/react-query-4` or `@suspensive/react-query-5` is installed.
