--- docs/suspensive.org/src/content/ko/docs/codemods/migrateSuspensiveReactQueryPackage.mdx
@@ -1,57 +1,0 @@
----
-description: '삭제된 @suspensive/react-query 패키지를 버전별 @suspensive/react-query-4 또는 @suspensive/react-query-5로 마이그레이션하는 코드모드.'
----
-
-import { Callout } from '@/components'
-
-# `@suspensive/react-query` 패키지 마이그레이션
-
-<Callout type='info'>
-
-- `@suspensive/react-query` v3 환경에서 `@suspensive/react-query` v4로 업데이트하는 경우 추천합니다.
-- Suspensive React Query에서 Deprecated 된 API를 TanStack Query에서 제공하는 API까지 변환하려면 [`tanstack-query-import`](./tanstackQueryImport) Codemod를 이용해주세요.
-
-</Callout>
-
-v4에서 통합 패키지인 `@suspensive/react-query`가 제거되었습니다. 설치된 `@tanstack/react-query` 메이저 버전에 맞춰 버전별 패키지(`@suspensive/react-query-4` 또는 `@suspensive/react-query-5`)로 변경해야 합니다.
-
-```bash filename="Terminal"
-npx @suspensive/codemods migrate-suspensive-react-query-package .
-```
-
-예:
-
-```tsx /@suspensive/react-query/
-import {
-  useSuspenseQuery,
-  SuspenseQuery,
-  queryOptions,
-  IsFetching,
-} from '@suspensive/react-query'
-import * as RQ from '@suspensive/react-query'
-export { queryOptions, QueryClientConsumer } from '@suspensive/react-query'
-```
-
-변환 후:
-
-```tsx /@tanstack/react-query/ /@suspensive/react-query-5/
-import { useSuspenseQuery, queryOptions } from '@tanstack/react-query'
-import { SuspenseQuery, IsFetching } from '@suspensive/react-query-5'
-import * as RQ from '@suspensive/react-query-5'
-export { queryOptions } from '@tanstack/react-query'
-export { QueryClientConsumer } from '@suspensive/react-query-5'
-```
-
-## 버전 감지
-
-현재 작업 디렉토리에서 위로 올라가며 가장 가까운 `package.json`을 확인해 `@tanstack/react-query`의 v4와 v5 중 어느 버전에 의존하는지 판단합니다. 감지된 메이저 버전에 따라 대상 패키지(`@suspensive/react-query-4` vs `@suspensive/react-query-5`)가 결정됩니다.
-
-감지에 실패한 경우 v5로 fallback 하며 경고를 출력합니다. 명시적으로 대상을 지정하려면 실행 전 `SUSPENSIVE_RQ_TARGET` 환경 변수를 `4` 또는 `5`로 설정합니다.
-
-```bash filename="Terminal"
-SUSPENSIVE_RQ_TARGET=4 npx @suspensive/codemods migrate-suspensive-react-query-package .
-```
-
-## 주의사항
-
-- `package.json`을 수정하지 않습니다. 실행 후 의존성에서 `@suspensive/react-query`를 제거하고 `@suspensive/react-query-4` 또는 `@suspensive/react-query-5`가 설치되어 있는지 확인하세요.
