--- packages/codemods/src/transforms/testfixtures/migrate-suspensive-react-query-package/migrate-suspensive-react-query-package.input.jsx
@@ -1,16 +1,0 @@
-import { useSuspenseQuery } from "@suspensive/react-query";
-import { SuspenseQuery } from "@suspensive/react-query";
-import { useSuspenseQueries, SuspenseQueries } from "@suspensive/react-query";
-import type { UseSuspenseInfiniteQueryOptions, IsFetching } from "@suspensive/react-query";
-import { useSuspenseInfiniteQuery as useSIQ, SuspenseInfiniteQuery as SIQ } from "@suspensive/react-query";
-import { usePrefetchQuery } from "@suspensive/react-query";
-import { PrefetchQuery } from "@suspensive/react-query";
-import { useQuery } from "@tanstack/react-query";
-import { unknownThing } from "@suspensive/react-query";
-import * as RQ from "@suspensive/react-query";
-import "@suspensive/react-query";
-import { SuspenseQuery as KeepSuspense } from "@suspensive/react-query-5";
-import { PrefetchInfiniteQuery, queryOptions as keepQueryOptions } from "@suspensive/react-query-4";
-export { Mutation } from "@suspensive/react-query-4";
-export { queryOptions, QueryClientConsumer } from "@suspensive/react-query";
-export * from "@suspensive/react-query";
