--- packages/codemods/src/transforms/testfixtures/migrate-suspensive-react-query-package/migrate-suspensive-react-query-package.output.jsx
@@ -1,20 +1,0 @@
-import { useSuspenseQuery } from "@tanstack/react-query";
-import { SuspenseQuery } from "@suspensive/react-query-5";
-import { SuspenseQueries } from "@suspensive/react-query-5";
-import { useSuspenseQueries } from "@tanstack/react-query";
-import type { IsFetching } from "@suspensive/react-query-5";
-import type { UseSuspenseInfiniteQueryOptions } from "@tanstack/react-query";
-import { SuspenseInfiniteQuery as SIQ } from "@suspensive/react-query-5";
-import { useSuspenseInfiniteQuery as useSIQ } from "@tanstack/react-query";
-import { usePrefetchQuery } from "@tanstack/react-query";
-import { PrefetchQuery } from "@suspensive/react-query-5";
-import { useQuery } from "@tanstack/react-query";
-import { unknownThing } from "@suspensive/react-query-5";
-import * as RQ from "@suspensive/react-query-5";
-import "@suspensive/react-query-5";
-import { SuspenseQuery as KeepSuspense } from "@suspensive/react-query-5";
-import { PrefetchInfiniteQuery, queryOptions as keepQueryOptions } from "@suspensive/react-query-4";
-export { Mutation } from "@suspensive/react-query-4";
-export { queryOptions } from "@tanstack/react-query";
-export { QueryClientConsumer } from "@suspensive/react-query-5";
-export * from "@suspensive/react-query-5";
