+++ packages/codemods/src/transforms/testfixtures/migrate-to-v4/migrate-to-v4.output.jsx
@@ -2,9 +2,16 @@ import { useSuspenseQuery, queryOptions } from "@tanstack/react-query";
 import { SuspenseQuery, IsFetching } from "@suspensive/react-query-5";
 import { mutationOptions, usePrefetchInfiniteQuery } from "@tanstack/react-query";
 import type { UseSuspenseInfiniteQueryOptions, SuspenseQueriesOptions } from "@tanstack/react-query";
-import { useSuspenseQuery as useSQ } from "@tanstack/react-query";
 import { PrefetchQuery } from "@suspensive/react-query-4";
+import { useSuspenseQuery as useSQ } from "@tanstack/react-query";
+import { PrefetchInfiniteQuery } from "@suspensive/react-query-4";
+import { queryOptions as keepQueryOptions } from "@tanstack/react-query";
+import { SuspenseQuery as KeepSuspense } from "@suspensive/react-query-5";
+import { unknownThing } from "@suspensive/react-query-5";
 import * as RQ from "@suspensive/react-query-5";
+import "@suspensive/react-query-5";
 import { useQuery } from "@tanstack/react-query";
 export { queryOptions } from "@tanstack/react-query";
 export { QueryClientConsumer } from "@suspensive/react-query-5";
+export { Mutation } from "@suspensive/react-query-4";
+export * from "@suspensive/react-query-5";
